requirements.txt
Django
djangorestframework
pytest
pytest-django
pytest-cov
mixer
mock
setuptools
factory_boy